ENH: Add load_user_token
Also: Refactored save_user_token to _persistence
